refactor:updated the code for fetching the movie geners
